chore(rfcs): Add RFC process documentation and template (#287)
Adding some more structure around how we propose, discuss and review changes to the public API, rather than waiting until a PR is opened, which is often a bit too late in the process. This has been adapted from Yarn's RFC template, which itself was based on Rust's RFC template.