feat(ScreenReaderOnly): support for different component types
ScreenReaderOnly support for different component types so that it can be used in different contexts.
ScreenReaderOnly is used as a child of both div and span elements with different child elements of
Block and Inline elements types. To be able to output valid HTML, choosing the component type can
allow this. By retaining default of rendering the children within a span element makes the change
backwards compatible.