docs(site): Configure router basename (#334)
Upgrading React Router seems to have broken the docs on GitHub Pages, even though they work on Surge. It seems likely to be caused by deprecation of transparent base href tag support, but regardless of whether this is the actual cause of the issue or not, we now need to explicitly provide the basename to our history instance.