feat(FieldLabel): Decouple FieldLabel styling with valid prop
Styling of the field label has previously been controlled by the valid prop. This can lead to
unwanted styling as the validity of the form does not always dictate the desired styling. For
example, a field message for a valid field should not always be green, this text could be used to
guide the user instead