fix(Autosuggest): Fix accessibility bug for screen readers (#526)
Current version of Autosuggest is unusable with screen readers. Bug has been introduced in v10.1.2 of react-autowhatever module. So we are pinning react-autowhatever (and also autosuggest itself) to user previous non-buggy version.

This PR will reduce accessibility for OSX due to rollback of the fix in v10.1.2, however it is still usable, thus more preferable.